requirements.txt
PyExifTool==0.1
python-magic==0.4.6
wheel==0.23.0
